chore: refact module qrcode & clean code
