+++ resources/js/components/app-sidebar.tsx
@@ -2,7 +2,7 @@ import { NavMain } from '@/components/nav-main';
 import { Sidebar, SidebarContent, SidebarHeader, SidebarMenu, SidebarMenuButton, SidebarMenuItem } from '@/components/ui/sidebar';
 import { type NavItem } from '@/types';
 import { Link, usePage } from '@inertiajs/react';
-import { ArrowBigDown, ArrowBigUp, Calendar, Dock, LayoutGrid, User, Users } from 'lucide-react';
+import { ArrowBigDown, ArrowBigUp, Calendar, Dock, LayoutGrid, QrCode, User, Users } from 'lucide-react';
 import AppLogo from './app-logo';
 
 type Role = 'Super Admin' | 'Admin' | 'Finance' | 'User' | 'Guest' | 'LimitedAccess';
@@ -45,8 +45,8 @@ export function AppSidebar() {
     },
     {
       title: 'Buat Absensi',
-      href: '/absensi/generate/qr-code',
-      icon: Calendar,
+      href: '/qr-code/create',
+      icon: QrCode,
     },
   ];
