+++ routes/web.php
@@ -1,12 +1,12 @@
 <?php
 
-use App\Http\Controllers\CreateQRCodeController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\KasController;
 use App\Http\Controllers\KegiatanController;
 use App\Http\Controllers\LaporanController;
 use App\Http\Controllers\PemasukanController;
 use App\Http\Controllers\PengeluaranController;
+use App\Http\Controllers\QRCodeController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
@@ -35,14 +35,15 @@ Route::middleware(['auth', 'verified', 'role:Super Admin'])->group(function () {
     Route::get('/users/export/csv', [UserController::class, 'exportCsv'])->name('users.export.csv');
     Route::resource('kegiatan', KegiatanController::class)->except(['index', 'show']);
     Route::get('/kegiatan/export/csv', [KegiatanController::class, 'exportCsv'])->name('kegiatan.export.csv');
+    Route::get('/qr-code/create', [QRCodeController::class, 'generate'])->name('qr.create');
+    Route::post('/qr-code/create', [QRCodeController::class, 'store'])->name('qr.store');
+    Route::post('/qr-code/{id}/deactivate', [QRCodeController::class, 'deactivate'])->name('qr.deactivate');
 });
 
 Route::middleware(['auth', 'verified'])->group(function () {
     Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
     Route::get('/kegiatan/terbaru', [KegiatanController::class, 'terbaru'])->name('kegiatan.terbaru');
     Route::get('/api/kegiatan/notification', [KegiatanController::class, 'notifications'])->name('kegiatan.notifications');
-    Route::match(['get', 'post'], '/absensi/generate/qr-code', [CreateQRCodeController::class, 'generate'])->name('qr.generate');
-    Route::post('/qr-code/{id}/deactivate', [CreateQRCodeController::class, 'deactivate'])->name('qr.deactivate');
 });
 
 require __DIR__ . '/public.php';
